Reduce the size of the Docker image
Introduces a multi-stage build, with a builder container that prepares the app
dependencies and a 'prod' container that only includes dependencies that are
required at runtime.

The size of the final container image is reduced by 60%. Smaller container
images reduce resource usage, make deployments faster and reduces the attack
surface of the app.